+++ src/card/WindRoseCard.ts
@@ -30,7 +30,7 @@ import {MeasurementHolder} from "../measurement-provider/MeasurementHolder";
 
 /* eslint no-console: 0 */
 console.info(
-    `%c  WINROSE-CARD  %c Version 2.0.0 `,
+    `%c  WINROSE-CARD  %c Version 2.2.0 `,
     'color: orange; font-weight: bold; background: black',
     'color: white; font-weight: bold; background: dimgray',
 );
@@ -247,7 +247,10 @@ export class WindRoseCard extends LitElement {
         if (this.cardConfig && this.updateInterval === undefined) {
 
             this.updateInterval = setInterval(
-                () => this.refreshMeasurements(!this.cardConfig.disableAnimations), this.cardConfig.refreshInterval * 1000);
+                () => {
+                    this.cardConfig.activePeriod.recalculateTimeRange();
+                    this.refreshMeasurements(!this.cardConfig.disableAnimations)
+                }, this.cardConfig.refreshInterval * 1000);
             this.log.info('Interval running with ' + this.cardConfig.refreshInterval + ' seconds.');
         }
     }